Apply to meta and data snapshots too
